Remove unneeded chef-server migration code
The vendored runit recipe of omnibus-chef-serfer includes one-time
migration code for installations that have a `opscode-runsvdir` service.
Due to unintelligent search-and-replace editing of yours truly, we not
only kept migration code we do not need (because there exist no previous
installations of omnibus-gitlab) but we were also triggering the cleanup
on files/services that should not be cleaned up. This commit removes the
migration code.