Defer LDAP defaults to upstream
Omnibus-gitlab should not have its own set of LDAP defaults. Either the
user adds a setting in /etc/gitlab/gitlab.rb, or gitlab-rails picks a
default value. Prior to this change, omnibus-gitlab would augment the
user settings in /etc/gitlab/gitlab.rb with its own set of defaults,
which is confusing and leads to configuration errors.