Merge branch 'master' into 'master'
Debian/Devuan issues on init system detection

- Devuan GNU/Linux uses sysvinit by default but can also use upstart
- Debian GNU/Linux uses systemd by default but can also use sysvinit or upstart
- Devuan GNU/Linux isn't supported in runit recipe for init

For this reason with the old way to detect based on distro/version mapping gitlab-ctl reconfigure will fail and stall in an infinite loop when
installing on Devuan or Debian with a non-default init system installed ( a legit config supported by both Devuan and Debian ).

This merge request also add support for Devuan in runit recipe for init.

See merge request !338