Use package-script erbs to include install_dir
Moved the package-scripts into the templates directory, and use a new package-scripts software to render them into the install directory. gitlab.rb was updated to additionally use that install directory for it's scripts, and exclude the directory from being included in the package source files.

The makeselfinst script was removed, as it appears that omnibus 4 was overwritting it with its own anyways.

Update the project path for omnibus-ctl

omnibus-ctl uses the passed project name to form the path information. So it is more accurate to use the install_dir rather than the project name when calling omnibus-ctl