Fixes an issue where the stat check we were using on storage diretories was english language dependant
The printf filetype from stat, returning 'directory' when in english, returns other things in when language is set differently. Added a -d expression to the test statement.

Technically stat has raw output that we could have masked the bit and parsed to determine if it is a directory, but this should suffice.