Merge branch 'cacert_update' into 'master'
Update to latest cacerts file

Tried to build a package this morning and it looks like the cacerts.pem file from curl.haxx.se has been changed recently.   Updated with a new version number and checksum.  

I have also updated the URL to get from https as if you try and get it with curl on a command line it redirects to https so probably better to define the download over https as well:

```
~> curl http://curl.haxx.se/ca/cacert.pem
<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML 2.0//EN">
<html><head>
<title>301 Moved Permanently</title>
</head><body>
<h1>Moved Permanently</h1>
<p>The document has moved <a href="https://curl.haxx.se/ca/cacert.pem">here</a>.</p>
<hr>
<address>Apache Server at curl.haxx.se Port 80</address>
</body></html>
```

I think this may be needed in 8.12.2 when that goes out, let me know if the CHANGELOG needs to corrected for that though.

See merge request !1007