Add gitaly RunIt service to GitLab Omnibus
The package does not include the executable itself at this time.

We intend to initially ship it seperatly, to allow it to iterate
in production on a seperate schedule from GitLab, at least initially.

But if you have the executable installed, and you enable the service,
It should now be run using RunIt and available during gitlab-ctl status/up/down/restart