Rename Trigger schedule to Pipeline Schedule
The feature was listed as experimental on the docs page describing this
feature, so we're free to change the key to match the worker name.

The frequency was improved too, so the peaks would be lower, and
accuracy would be better. The default cron pattern was choosen so it
would spread the load between other scheduled jobs nicely.