Fix documentation on gitlab_monitor listen_address
gitlab_monitor requires address and port as separate settings, as they
are templated as separate settings in the `gitlab-monitor.yaml`.

The example as it was before this change cannot work, WeBrick would try
to resolve the passed address in `listen_address` and fail to due so due
to the added port - which results in gitlab_monitor failing.