Docker: update-permissions, only touch registry if exists
Add an exists check to the `update-permissions` script, so that it only attempts to set the permissions on the registry (`/var/opt/gitlab/gitlab-rails/shared/registry`) if it actually exists.

This removes the spurious "fail" on setting permissions on things that don't exist, and allows for better scripting compatability.

Fixes #2074