Fix gitlab-shell not able to import projects from hosts with trusted SSL certificates
The git binary by default uses libcurl's CURL_CA_PATH to search for
certificate authorities. By default, libcurl autodetects the system paths
instead of using the omnibus-gitlab path:
https://github.com/curl/curl/blob/curl-7_54_1/acinclude.m4#L2640-L2645

This change specifies the CA path to libcurl so that curl and git will work
out of the box with the trusted certificates.

Closes #2180