Attempt to use a new Mash object that auto-vivifies (auto-creates) during the gitlab.rb parsing.
It's called vivify because that's what Chef calls it in their VividMash,
and that's where we are using the [] and conver_value methods from to
create ours.

It really just means it auto-creates a new mash when you ask for a key
that doesn't yet exist.

We don't want this on all the time, because we often check these keys as
booleans to determine if they are set (and expect them to return nil)

So we only auto-vivify during the from_file call.