Fix an issue where we would link our binaries to the system binaries
This would occur when an old data dir containing a PG_VERSION that was
no longer shipped with GitLab was found.

The fix now falls back to the psql version, if the PG_VERSION file is
present, but the matching binary isn't shipped.

And introduces an additional error if, for some reason, the psql version
check fails or doesn't match our postgres version.