Remove ignored/misleading environment variable
GO15VENDOREXPERIMENT was introduced in go 1.5 defaulting to off.  In 1.6
it defaults to on, and from 1.7 onwards it just the default behavior so
it's not even consulted.[1]

Our build machines currently use 1.8.1 (gitlab-omnibus-builder!40) so
there is no need to specify GO15VENDOREXPERIMENT.  Also this allows us
to prune out a misleading comment.

[1] https://github.com/golang/go/wiki/PackageManagementTools#go15vendorexperiment