Add roles for postgres HA
Here we add the `postgres`, `pgbouncer`, and `consul` roles, and add a
quick declaration to the documentation that there roles exist.

As a style issue, we add supporting service groups, named for the roles
that use them.  This should make extending the roles more obvious.

In the postgres role we continue to have to enable repmgr in addition to
its service group as it has multiple settings
(`Gitlab['repmgr']['enable']` and `Gitlab['repmgrd']['enable']`) that
need to be true to correctly enable repmgrd.