Mattermost: Don't store GitLabSettings.*Endpoint in gitlab-secrets.yml
Because we stashed `GitLabSettings.*Endpoint` with the rest of the MM secrets in `gitlab-secrets.yml`, changing `external_url` would not result in a change to the `*Endpoint` values.

Remove these values from being stored in `gitlab-secrets.yml`

I am unsure if additional work is needed to automatically ignore these settings if present in `gitlab-secrets.yml`, or if manual intervention will be required by the user.