Correct permissions of omniauth-jwt
omniauth-jwt 0.0.2 installs `omniauth/strategies/jwt.rb` with mode 0600.

This is probably correct on the authors machine, but prevents us from
using it reliably as we install files as the `root` user but run them as
the `git` user.

Here we add a targeted chown to ensure the files are readable every time.