Geo: Set recovery_target_timeline to latest by default
We needed this setting when a failover happen and the secondary wasn't
able to pull the latest segment from the current timeline. Setting to
latest allows the secondary to recovery in case this happens.

See https://www.postgresql.org/docs/9.6/static/continuous-archiving.html#BACKUP-TIMELINES
for more details.

Relates to gitlab-com/migration#407