Enable rbtrace for unicorn if ENABLE_RBTRACE is set
rbtrace was added in https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/20831, but
it only worked for Sidekiq because the unicorn master
resets any signal handlers that are set during the Rails initializers.
Moving the `require 'rbtrace'` calls to after the fork ensures that
the rbtrace listener works for unicorn.

Relates to https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/20892 as well.