Skip calling the preinstall script from the docker wrapper
Prior to 11.0, it appears to me that the only thing this accomplished
was creating the /var/log/gitlab/reconfigure directory. (The db backup,
posgres version checking, and upgrade version checking would not run
because the services folders is emptied before this runs, and any old
files that could be used to check the old version aren't present in the
image)

In 11.0 the script also added a config-check. This check normally
wouldn't do anything when upgrading an docker image, because the config
file it checks won't exist yet... But because we call the script on
every startup of the image here in Docker, it causes the check to be run
on every restart, even when its not an upgrade.

This causes issues, because if you ever have an error as a result of a incorrect
reconfigure, the image won't start, which makes it pretty hard
to fix.