Add a new pg_version helper to deal with major-minor version changes
This currently should change any behaviour, but is prep for two upcoming
changes.

One where move the install directory for postgres to only use the major
postgres version in the filepath.

And two, the introduction of Postgres 10, where it will matter that the
major version we are using of postgres is 9.6, rather than just 9,
during the transition. And that the version scheme changes a bit
starting in 10. https://www.postgresql.org/support/versioning/