Fix Docker registry not working with Windows layers
Docker v2.7.0 changed the behavior from v2.6.2 in that it enabled the
validations to check manifest URLs by default, which caused the registry
to set the allowed URL regexp to nothing (`^$`) instead of `nil`. This
caused the container registry to reject any foreign layers that
contained URLs. We fix this by disabling the validations with the
newly-introduced `disabled` flag.

See https://github.com/docker/distribution/issues/2795 for more details.

Closes https://gitlab.com/gitlab-org/omnibus-gitlab/issues/4013