Support Google Cloud Memorystore by disabling Redis CLIENT
MemoryStore is a Redis instance that has disabled commands such as
CLIENT and MONITOR. Sidekiq calls CLIENT to set an ID for debugging, but
this can be disabled by setting the `id` to `nil`
(https://github.com/mperham/sidekiq/issues/3521).

This is done in `/etc/gitlab/gitlab.rb` via:

gitlab_rails['redis_enable_client'] = false

By default, this is set to `true`.

Closes https://gitlab.com/gitlab-org/omnibus-gitlab/issues/4149