Stop printing Skipped message in the trusted certificate helper
When there are issues with user's certs, they often get confused by
these messages, as they print for all files in the folder that aren't
symlinks, including valid certficates.

To be copied at this stage, the cert needs to be valid, and be a
symlink. It's filename a hash. Rather than print what was skipped
(which includes the valid targets of the symlinks). Let's just print
what was linked over, and let the absense of a message work better for
debugging.