Praefect service in Omnibus
For HA Git storage, the Gitaly team is building a daemon to manage all
Git data, called praefect. This daemon needs to be build, which is done
by the Makefil included in the repository.

The configuration part was missing however, and this change adds it. The
service is disabled by default, as it's experimental for now.