Fix SELinux installation failures on CentOS 8
CentOS 8 requires that pathnames be present when calling `restorecon`,
but this was not a requirement in CentOS 7. We fix this by moving all
SELinux-related changes to the gitlab::selinux recipe so that all the required
files and directories can be made before any SELinux changes are applied. This
makes sense from a code organization standpoint and also has the nice side
effect of removing the need to ignore unknown files (-i option in restorecon)
as well.

This commit is essentially commit 5b99224684f49a19234b22fb2ed49ba814a355be by
Stan Hu, rebased against latest master

Signed-off-by: Balasankar "Balu" C <balasankar@gitlab.com>