Fix inability to disable Connection and Upgrade NGINX headers
Since the `proxy_set_headers` are defined as node attributes, Chef does
a deep merge and makes it impossible to disable them. For example,
when `{ "Connection" => "$upgrade" }` gets merged with
`{ "Connection" => nil }`, the value in the node attributes wins,
resulting in `{ "Connection" => "$upgrade" }`.

To fix this, if the user nulls out the header, we delete the header from
the default attribute directly.

Closes https://gitlab.com/gitlab-org/omnibus-gitlab/issues/4873