Add max_request_duration config to gitlab_rails
This option lets the rails app know how much time it has to process a
request. And make decisions accordingly.

By default the time is 90% of the timeout configured for the
worker. This gives us some time during the request to clean up if we
need to.

Setting a duration higher than the worker timeout will raise an error.