Delete unused code to find and delete earlier AMIs
Initially we wanted to only have the latest AMI available at any point
of time. So we wrote the code to delete the earlier images, but marked
it as dry run to ensure it did its job without actually touching
anything. But we never removed that dry run flag.

Also, since we have customers who may intentionally remain on older
versions, we do have to maintain older images. So there is no need for
this code anymore.