Disable RubyGems for Gitaly gitlab-shell hooks
By default, RubyGems is enabled, which causes the Ruby interpreter to
load 600+ additional files at startup, even though gitlab-shell is
designed not to use any external gems.

This was done in
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/1707/diffs
for gitlab-shell, but when the hooks migrated to Gitaly we lost the
performance optimization.

Also remove gitlab-shell wrapper since it's no longer needed.