Praefect: Enable SQL failover by default
Now Praefect has the ability to disable mutation on repositories when
these are out of date and failover happens, data loss should not happen
anymore.

As such, this will now be the default, after: https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2218

This change updates the omnibus defaults to reflect this, but is not
blocked on the upstream merge request.