Manually disable copy_file_range() on RedHat kernels
Calls to the `copy_file_range()` system call on recent RedHat kernels
now fail with `EOPNOTSUP` instead of `ENOSYS`, which breaks any use of
`IO.copy_stream` (e.g. used by CarrierWave) if NFS is used.

Ruby attempts to detect whether the kernel supports this system call,
but the detection mechanism only checks whether this system call is
defined, not whether it is actually supported. We manually patch this to
disable the `copy_file_range()`.

Relates to:
1. https://bugs.ruby-lang.org/issues/16965
2. https://gitlab.com/gitlab-org/gitlab/-/issues/218999
3. https://bugzilla.redhat.com/show_bug.cgi?id=1783554