Replace regexp checks with config equality checkin Praefect specs
This commit replaces a bunch of ad hoc regexp checks in Praefect's
config file spec with a single equality assertion. This covers more
ground as any new fields will raise an error if not added to the
assertion. This also eases maintenance as the whole config file is
easily asserted against.