Adds an additional inlude_ee? check to prep for additional packages
The idea is to have packages that include the ee source and license, but
aren't specifically the gitlab-ee package. (How those would work, is
that they would be added as additional checks in inlude_ee?)

A cleaner way might be some metadata info that has keys indexed by package
name that indicates whether a package is supposed to include ee or not.
But I'm not sure its worth the effort at this point.