Add net-protocol BSD-2 license
As announced in
https://www.ruby-lang.org/en/news/2020/12/25/ruby-3-0-0-released/,
net-http and net-protocol have been promoted from the Ruby stdlib to a
default gem. The update to mail_room in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/56913 will depend
on a newer version of net-imap and net-http, which in turn depends on
this newly-released net-protocol gem.

These default gems are all released under the Simplified BSD license, so
we should be able to use them.

Changelog: added