Deprecate configuring Gitaly nodes in virtual storage's config root
Historically, a virtual storage's Gitaly nodes have been configured root
of the virtual storage's config object. This didn't work well, as the
configuration format did not match the structure Praefect itself was using
and made it awkward to add virtual storage scoped configuration keys. Since
then, we've modified the configuration format in Omnibus to also put the
nodes under the 'nodes' key and included fallback code to retain backwards
compatibility. In 15.0, we have the chance to remove the fallback code. To
prepare for that, this commit adds logs a deprecation message if the there
are some nodes configured in the root of the virtual storage's config rather
than under the 'nodes' key.

Changelog: deprecated