Adds default for prepared_statements config in Geo secondary db.yml
Rails is expecting prepared_statements to be either true or false,
and prior to Rails 6.1, invalid values were "converted" to false,
but in Rails 6.1, the interpreted value is either "false", or what
is passed in the config literally.

This adds a default value set to false, exactly as for the normal
database, to prevent the invalid value.

Changelog: fixed