Update Praefect defaults in template
The praefect settings in `gitlab.rb` template were getting out of date.

This commit updates removes `praefect['failover_election_strategy']`
and `praefect['virtual_storage_name']`, which are no longer in use. In
addition, it updates `praefect['failover_enabled']` to `true`, as this
is the default setting for this value.

Changelog: removed