Centralize enabling and disabling of crond
Prior to GitLab 13.5, crond was only needed for auto-renewing
LetsEncrypt certificates. If the auto-renew functionality were disabled,
the LetsEncrypt recipe would deactivate crond.

GitLab 13.5 introduced an automatic database reindexing task via
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/4602. This
removed the automatic disabling of crond.

As a result, users upgrading prior to GitLab 13.2 may find that the
crond remains activated with a stale config file, preventing the process
from starting up due to argument errors. This config file would only be
updated if one of the above features were activated.

To fix this issue, we centralize the enabling and disabling of crond in
the default recipe and remove this logic from the other recipes.

Relates to https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/5541

Changelog: fixed