Nginx: modernise TLS config
This commit updates updates the TLS settings to the intermediate
settings from the Mozilla SSL Configuration Generator [1]

The preference for server ciphers is disabled and more modern ciphers are
chosen.  The builtin session cache is not set explicitly anymore, because the
value was smaller than the default.

HSTS times are bumped to two years instead of one.

This is a companion MR to gitlab-org/gitlab!65345

[1] https://ssl-config.mozilla.org/

Changelog: changed