Nginx: further modernisation and standardisation
This commit puts in the remaining pieces to updates the TLS settings to the
intermediate settings from the Mozilla SSL Configuration Generator [1]

session tickets are disabled, since the timeout is increased and renegotiating a
session should be cheap enough. Sections for HSTS are added were missing.

This is a companion MR to gitlab-org/gitlab!65345

[1] https://ssl-config.mozilla.org/

Changelog: added