Connect Puma low-level error handler to Sentry
We saw in
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/5194 that
errors in a middleware would only be caught by Puma's low-level handler.
This error only showed up in the Puma stderr log, which made it
difficult to diagnose.

We now send the exception to Sentry if it is active. The exception will
still be printed in stderr as before.

Closes https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/6284

Changelog: added