Drop --without-gnutls flag from curl
https://github.com/curl/curl/commit/68d89f242cf9f6326e3b2f6fe119b7c74ef41c66
dropped the use of `--without-gnutls` in favor of `--with-*`
options. This commit actually had the opposite effect of enabling GnuTLS
when `--without-gnutls` were set. https://github.com/curl/curl/pull/7994
fixes this behavior.

We can drop this now that this does completely the opposite of what it
is supposed to be doing.

Discovered in https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/5592#note_731305703

Changelog: fixed