Include the cert file location by default as well
Setting only the cert directory will cause some ruby gems to not load
the default paths for the openssl store. We can't only use the cert
directory because we don't provide hashes for all the cacerts in
omnibus. Omnibus relies on the cacerts being provided from the ca cert
file.

So for ruby processes we should explicitly include it.

Changelog: fixed