Enable frame pointer in PostgreSQL compile options
As we did with
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/4030 and
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/3421,
enabling the frame pointer is critical to being able to use `perf` with
PostgreSQL.

This is enabled in the PostgreSQL PGDG builds
(https://wiki.postgresql.org/wiki/Apt). I was not able to get a
flamegraph from the Omnibus PostgreSQL build in
https://gitlab.com/gitlab-org/gitaly/-/issues/3973, so I ended up
switching to a PGDG build.

Changelog: changed