gitaly: Install both external and bundled Git
The Gitaly project is in the process of migrating to bundled Git, which
allows it to gain tighter control over which Git versions are installed,
allowing for the use of feature flags to roll out new Git versions. This
will eventually replace the Git distribution which we're installing
nowadays, and thus also provide space savings given that we only need to
install a subset of the Git distribution.

Prepare for this migration by installing both external and bundled Git.

Changelog: changed