spec: Fix test for rendering default Git binary path
When we've introduced a default Git binary path in b157bbe56 (gitaly:
enable bundled git as default and set a default binary, 2022-02-18), we
also added a spec which asserts that we write the expected default Git
binary path. But this spec doesn't make a lot of sense: we assert that
we _don't_ write the default path if unset, which only passed because we
were accidentally missing the quotation marks. Instead, we should be
asserting that the default path is written as expected.

Fix the spec to do so.