Expose `db_database_tasks` attribute
This attribute configures if database should be managed
by `db.rake` tasks. This is not desired for additional databases
if they point to exactly the same `main:` database
as this causes a troubles when provisioning or migrating.

The default is `database_tasks: true`, but if Omnibus detect
that host/port/database equals to main it will automatically
set it to `false`.

If the host is not equal, but still points to the same database
admin should configure `database_tasks: false` by themselves.

Changelog: added