consul: Disable logging of timestamps if JSON log enabled
Consul has a `log_json` parameter that will log the timestamp.
Previously runit would supply the `-tt` option, which would always
preface the log line with its own timestamp. This would make it harder
to ingest the logs since the generated logs are not in
newline-delimited JSON format, so we now omit this extra timestamp.

Changelog: fixed