Only enable Go related license checks for gitlab-pages
The gitlab-pages upstream recently introduced dangerbot. The
gems are not in the production group, but the build process
never actually runs bundler because gitlab-pages is written
in pure Go. This caused license-finder to fail repeatedly
attempting to check the licenses for the dangerbot gem when
it was not actually part of the build at all. This configures
license-finder to only check licenses for Go dependencies
when building gitlab-pages.

Related https://gitlab.com/gitlab-com/gl-infra/production/-/issues/6841