gitaly: Remove configuration for Rugged's gitconfig search path
Gitaly is still using Rugged in its Ruby sidecar to provide a very
limited set of RPCs that have not yet been ported to Go. While Rugged in
the past required the gitconfig file to function correctly, nowadays it
only requires the `core.fsyncObjectFiles` config entry. This entry must
be set to `true` or otherwise it is trivial to end up with repository
corruption.

This means that the gitconfig must be in a specific well-known state for
Rugged to operate correctly. But currently there is still the option to
set Rugged's gitconfig search path, which must in fact point to a path
that contains a gitconfig with the mentioned config entry. This is quite
fragile without any benefit for the administrator at all.

To fix this, Gitaly has introduced a fallback mode via [1] that causes
it to write a Rugged-specific gitconfig file in case the Rugged search
path is not set up. This allows us to now phase out support for setting
this configuration altogether: it's not required, may cause repository
corruption if misconfigured, and ultimately only makes the architecture
harder to reason about given that we currently share this gitconfig file
between multiple components.

Remove the ability to set Rugged's gitconfig search path without any
replacement.

[1]: https://gitlab.com/gitlab-org/gitaly/-/merge_requests/4560

Changelog: removed