Skip auto-restart of PG during reconfigure as part of pg-upgrade
In nodes where Geo PG runs along with regular PG, during the PG upgrade
process of regular PG, a reconfigure is initiated. This reconfigure will
update the `VERSION` file in Geo PG's directory also with the new
verison information and trigger a restart of the Geo PG. However,
because Geo PG's data is still of the old version but the binaries have
already been updated to point to the new version, this restart will fail
due to mismatch.

Because PG upgrade anyways causes the DB to be stopped and started, we
can safely skip the automatic restart as part of reconfigure and avoid
this failure.

Changelog: fixed
Signed-off-by: Balasankar "Balu" C <balasankar@gitlab.com>