Reinstall Ruby grpc gem with system OpenSSL for FIPS builds
By default, the Ruby grpc gem installs itself as a native gem and
bundles its own copy of BoringSSL. BoringSSL is a fork of OpenSSL, and
while there are FIPS-validated versions available, a BoringSSL FIPS
library is not generally available. However, we can ensure FIPS
compliance by building and linking against the system OpenSSL.

Unfortunately, the gRPC upstream has not made it possible to configure
the Ruby compilation options
(https://github.com/grpc/grpc/pull/27881).

To work around this, we patch the Ruby `extconf.rb` to link against
the system OpenSSL. This is done via a separate `ruby-grpc` software
Omnibus component. We do this for two reasons:

1. Since both GitLab Rails and Gitaly attempt to install the grpc gem,
it's possible for `bundle install` to install both the Ruby and the
native platform versions at the same time.

2. There is also the possibility that different gem versions get
installed by different components, such as GitLab Rails and Gitaly,
and we should allow for that.

This commit:

1. Determines which grpc gems have been installed in the system.
2. Uninstalls all versions.
3. For each version, download the gem, patch it with the handy
   `gem-patch` tool, and reinstall the patched gem.

On my local test, the rebuild of the gem added 150 seconds to the
build time.

Relates to https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/6687